+++ game_patch/object/entity.cpp
@@ -264,19 +264,21 @@ FunHook<void(int)> entity_blood_throw_gibs_hook{
 
         // skip entities with ambient flag (is in original but maybe not necessary?)
         rf::Entity* entity = (objp->type == rf::OT_ENTITY) ? static_cast<rf::Entity*>(objp) : nullptr;
-        if (entity && (entity->info->flags & 0x800000)) {
-            return;
+        if (entity) {
+            if (entity->info->flags & 0x800000) {
+                return;
+            }
+
+            // delete entity muzzle light if active, prevents persistent dynamic lights after entity is deleted
+            if (entity->muzzle_light_handle > -1) {
+                rf::gr::light_delete(entity->muzzle_light_handle, 0);
+            }
         }
 
         // skip corpses that shouldn't explode (drools_slime or custom_state_anim)
         rf::Corpse* corpse = (objp->type == rf::OT_CORPSE) ? static_cast<rf::Corpse*>(objp) : nullptr;
         if (corpse && (corpse->corpse_flags & 0x400 || corpse->corpse_flags & 0x4)) {
             return;
-        }
-
-        // delete entity muzzle light if active, prevents persistent dynamic lights after entity is deleted
-        if (entity->muzzle_light_handle > -1) {
-            rf::gr::light_delete(entity->muzzle_light_handle, 0);
         }
 
         static constexpr float spin_scale_min = 10.0f;