+++ launcher_common/UpdateChecker.cpp
@@ -10,7 +10,8 @@
 // 4 = 1.2.0 (Willow)
 // 5 = unused
 // 6 = 1.2.1 (Willow)
-// 7 = 1.3.0 (TBD)
+// 7 = 1.2.2 (Willow)
+// 8 = 1.3.0 (TBD)
 
 void UpdateChecker::CheckForUpdates()
 {
@@ -19,7 +20,7 @@ void UpdateChecker::CheckForUpdates()
     xlog::info("Checking for updates...");
 
     // Create an HttpSession
-    HttpSession session("Alpine Faction v1.3.0 Update");
+    HttpSession session("Alpine Faction v1.2.2 Update");
 
     try {
         HttpRequest req(update_url, "GET", session);