+++ docs/CHANGELOG.md
@@ -13,6 +13,7 @@ Version 1.3.0 (TBD): Not yet released
 ### Bug fixes
 [@GooberRF](https://github.com/GooberRF)
 - Fix `Blue team` and `Red team` labels being swapped in `Scope_Gate` int1 field in level editor
+- Fix crash when gibbing some corpses in single player
 
 Version 1.2.1 (Willow): Released Dec-30-2025
 --------------------------------